% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot_2.R
\name{myboot_2}
\alias{myboot_2}
\title{My Bootstrap Function}
\usage{
myboot_2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{x, iter, fun, alpha}{The number of samples, number of iterations, function, and significance level.}
}
\value{
A histogram of the point estimation and the interval of the population value.
}
\description{
My Bootstrap Function
}
