% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybootpval.R
\name{mybootpval}
\alias{mybootpval}
\title{My Bootstrap P-Value Function}
\usage{
mybootpval(x, conf.level = 0.95, iter = 3000, mu0 = 0, test = "two")
}
\arguments{
\item{x, conf.level, iter, mu, test}{the dataset, the confidence level, iterations, the mean, and the style of test.}
}
\value{
Several vectors and a histogram.
}
\description{
My Bootstrap P-Value Function
}
