% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{My CLT Poisson Function}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n, iter, lambda}{vectors of quantitative data}
}
\value{
Graphs of CLT poisson distributions.
}
\description{
My CLT Poisson Function
}
