% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymax_lik}
\alias{mymax_lik}
\title{My Max Likelihood Function}
\usage{
mymax_lik(lfun, x, param, ...)
}
\arguments{
\item{lfun, x, param}{The function, the variable, and the parameters}
}
\value{
A graph and data describing the maximum likelihood of results.
}
\description{
My Max Likelihood Function
}
