% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mynbin.R
\name{mynbin}
\alias{mynbin}
\title{Negative Binomial Function}
\usage{
mynbin(y, r, p)
}
\arguments{
\item{y, r, p}{vectors of quantitative data}
}
\value{
a vector
}
\description{
Negative Binomial Function
}
\examples{
mynbin(10,3,0.5)
}
