% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myntickets.R
\name{myntickets}
\alias{myntickets}
\title{My N Tickets}
\usage{
myntickets(N, gamma, p)
}
\arguments{
\item{N}{number of seats on the flight.}

\item{gamma}{probability a flight will be overbooked.}

\item{p}{probability of a 'show'.}
}
\value{
a list and a plot of the objective function.
}
\description{
My N Tickets
}
